Simplify parsing of numbers
Instead of storing the original string in ElemNumVal, it now stores them
as parsed numbers, and numbers are validated when lexing the ABNF,
rather than on Parse(). This simplifies things quite a bit.

There shouldn't be any functional changes but the Fuzz_eaa469604868c87f
failed:

    --- FAIL: Test_U_ParseABNF (0.06s)
        --- FAIL: Test_U_ParseABNF/Fuzz_eaa469604868c87f (0.00s)
            grammar_test.go:209: Expected err: false ; got: strconv.ParseUint: parsing "111000000110110": value out of range

But I think that was actually a bug before? 111000000110110 is 28726, so
that seems like a correct error?